+++ src/main/java/com/equip/equiprental/common/exception/GlobalExceptionHandler.java
@@ -2,7 +2,7 @@ package com.equip.equiprental.common.exception;
 
 import com.equip.equiprental.common.controller.ResponseController;
 import com.equip.equiprental.common.dto.ResponseDto;
-import com.equip.equiprental.interceptor.RequestTraceIdInterceptor;
+import com.equip.equiprental.common.interceptor.RequestTraceIdInterceptor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
